Remove redirect
this redirect will never get hit.
The 'sessions#create' action will only be hit
from the root_path.
When the action is hit, the user goes to a Twitter auth page;
If the Twitter auth is successful, the  will redirect to the feed;
If the Twitter auth is unsuccessful,
the Twitter middleware page won't let the user proceed.